Update AccordionCustomHeaderContent.md
throws error when content and title are not explicitly password as params in _renderContent and _renderHeader methods respectively

error is : Invariant Violation: Invariant Violation: Objects are not valid as a React child (found: object with keys {title,content}). If you meant to render a collection of children, use an array instead.