Typo - update Accordion.md
The sample graphic for the "Accordion" component indicates that "arrow-down" is for closed and "arrow-up" is for open, which is also logical behavior. The docs indicate the opposite.

Note, I have not used NativeBase. I was looking through the docs while considering to use it and noticed what appears to be this typo.